refactor: Way의 패키지 경로 변경 나중 리펙터링을 위해 filter 패키지 생성
